requirements.txt
torch
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cu121

fastapi
uvicorn
python-multipart

faster-whisper
ctranslate2

requests
python-dotenv
ffmpeg-python
pydantic